web: guard audited game logic edge cases
Add boundary checks and type-safe restores for audited game logic paths that can fail under TeaVM or edge-map state, including traps, plants, scene inspection, tile refresh, and item or mob restoration.

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Monk.java
@@ -137,7 +137,7 @@ public class Monk extends Mob {
 	@Override
 	public void restoreFromBundle( Bundle bundle ) {
 		super.restoreFromBundle( bundle );
-		focusCooldown = bundle.getInt( FOCUS_COOLDOWN );
+		focusCooldown = bundle.getFloat( FOCUS_COOLDOWN );
 	}
 	
 	public static class Focus extends Buff {
